+++ tests/_routes-contract-test/echoes.php
@@ -1,0 +1,4 @@
+<?php
+// Invalid: direct output from a route file must throw
+echo "this is not allowed";
+return new \mini\Http\Message\Response('unreachable');