+++ tests/_routes-wildcard-test/users/_/friendship/_.php
@@ -1,9 +1,7 @@
 <?php
 // Test double wildcard: matches /users/{user_id}/friendship/{friend_id}
-header('Content-Type: application/json');
-echo json_encode([
+return new \mini\Http\Message\Response(json_encode([
     'handler' => 'users/_/friendship/_.php',
     'user_id' => $_GET[0] ?? null,
     'friend_id' => $_GET[1] ?? null,
-]);
-return null;
+]), ['Content-Type' => 'application/json']);