+++ tests/_routes-wildcard-test/users/_/posts.php
@@ -1,8 +1,6 @@
 <?php
 // Test nested wildcard: matches /users/{user_id}/posts
-header('Content-Type: application/json');
-echo json_encode([
+return new \mini\Http\Message\Response(json_encode([
     'handler' => 'users/_/posts.php',
     'user_id' => $_GET[0] ?? null,
-]);
-return null;
+]), ['Content-Type' => 'application/json']);